Test the issues with linestring geometry and stop distance from closest rawnav point
@wtimmerman-fitp : I checked the geometry of the nearest_rawnav_wmata_schedule_all_routes_days: it is linestring for all rows.

I plotted the trajectories for H8, the issue of large distance b/w stop and rawnav points is because rawnav trajectory goes over the sidewalk sometimes. Also, some of the stops in wmata schedule are marked away from the pavement thus distance of the closest rawnav point to the stop is large.